Changed autoId from int to long (missed one spot). Have now completed this change. (I think I really got them all now.)

+++ app/src/main/java/albright/csit/legosetstracker/LegoTheme.java
@@ -42,11 +42,11 @@ public class LegoTheme {
   public LegoTheme() {
   }
 
-  public LegoTheme(int themeId) {
+  public LegoTheme(long themeId) {
     this.autoId = themeId;
   }
 
-  public LegoTheme(int themeId, String themeName) {
+  public LegoTheme(long themeId, String themeName) {
     this.autoId = themeId;
     this.name = themeName;
   }
